Implemented Router State
Implemented a router state so that the same database instance can be used across the entire router.
Makes it so that I don't need to make a new database each time a router method is called.
Much more efficient but also needed to add a panic since if the database can't be created the API can't function.
Problem which will need to be fixed in the future.