Fixed issue with missing uuid field
Changed Vecs of table entries to be Things instead of the specific struct.
This means that only the table and id are saved instead of a copy of the struct.
This solves both the issue of infinite recursion between events and vendors, as well as the missing uuid bug that was causing runtime failures.
While this does solve the above bugs, it also means that you can no longer get an entry directly from its linked record. Instead you must query the database a second time
However due mainly to the fact that there shouldn't be many queries where you need to chain them together as mentioned, the overall performance hit should be negligable.