Add and Remove Routes
Finished add and remove routes for Event, Menu, and Items.
There are a few issues I still need to address, mainly a filter is missing and there is a overall lack of pre-validation on the json post body.
Additionally I still need to implement authentication to ensure arbitrary REST calls dont leak data/perform unvalidated actions on the db.
There was another note, but its late and I forgot. Guess it can't have been that important.