Start on auth middleware
Started working on auth middleware.
Moving to a new branch since this may not be the ideal way to accomplish this.
Using a simple function and axum::middleware::from_fn may be a better idea since it will take way less work and acomplish the same thing.
May want to pivot back in the future so leaving here.
This branch will likely become very stale and be deleted in about 2 months though.