Updated models to allow for delayed vendor linking
Added an option to create a struct with a vendor.
This allows for creating a struct from a json value (which does or doesn't include a "vendor_id" property) and then specifying the vendor_id with the .with_vendor function.
Current use case is when creating entries from post paths, the vendor id is in the uri while the data is in the body. By using a delayed vendor linking system it stops repetition of the id in the json and uri.
Additionally since linking happens in the post handler anyways the vendor id is not needed when the object is created, only when it is added to the database.