Added Patch Handlers
Added patch handlers for all four tables. Allows for patching of modifiable information. Information is passed in the json body and is updated one at a time.
This updating method was chosen in part due to it allowing one malformed patch to not stop the entire patch from occuring. However it was mainly chosen due to it being seemingly impossible to patch in bulk when you have optional arguments.
This method of patching is not ideal since it can cause many requests to be run against the database. So a better solution should be found.
Additionally changed the println err format from {msg}:{line} to {msg}. Found the line numbers werent helpful since they ended up moving quite frequently to different lines.