Frontend OAuth2
Added basic oauth2 integration to the frontend.
Frontend now includes a webpage with a button which prompts the user to login via facebook oauth2.
If the user fails to login, an error is displayed. If login is successful, a logout button is displayed.
Frontend does not work with old backend since auth route is not reflected in the routes provided by the current backend.